server/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
flask-restful
python-dotenv
gunicorn
passlib